+++ F1Manager2024Plugin/F1Manager2024PluginSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace F1Manager2024Plugin
+{
+    public class F1Manager2024PluginSettings
+    {
+        public string Path { get; set; } = null;
+    }
+}